ISSUE-191 Client: Moved currentQaId and qaRemaining states to ChallengeManager from ChallengeList.
